Fix TXCREATE Gas Costing issues
* Charge for Initcode in Initcode Transactions as part of intrinsic cost
* In TXCREATE opcode do not charge for initcode size nor initcode
hashing (handled in intrinsic gas)

+++ evm/src/main/java/org/hyperledger/besu/evm/gascalculator/IstanbulGasCalculator.java
@@ -42,7 +42,7 @@ public class IstanbulGasCalculator extends PetersburgGasCalculator {
   public IstanbulGasCalculator() {}
 
   @Override
-  protected long callDataCost(final long payloadSize, final long zeroBytes) {
+  protected long callDataCost(final int payloadSize, final int zeroBytes) {
     return clampedAdd(
         ISTANBUL_TX_DATA_NON_ZERO_COST * (payloadSize - zeroBytes), TX_DATA_ZERO_COST * zeroBytes);
   }
